demeter/requirements.txt
Flask
python-telegram-bot
python-dotenv
opencv-python-headless
requests
Pillow
litellm
chromadb==0.4.24
numpy<2.0.0
python-telegram-bot[job-queue]
langchain-community
langchain-core
langchain-chroma
pytz
schedule
gunicorn
